+++ Changelog.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 3.3.0
+
+**This release includes resizable and scrollable tooltips!**
+
+### New Features
+- Tooltips can now be resized to a percentage of the screen (max height and width percentage can be customized)
+- Tooltips that are too long to be displayed on the screen become scrollable: use your mouse wheel to scroll through them
+- There's now a new config screen accessible through the main menu by clicking the `Scroll & Custom Tooltip Dimensions` button! (Resize arrows icon)
+
+### Technical Fixes
+- Removed the unused CustomTooltipApitest class from the final jar file
+
 ## 3.2.0
 **This release brings full compatibility with both 1.20.1 Fabric and Forge, 1.21 Fabric and NeoForge as well as Spanish translations!**